+++ lukefi/metsi/domain/utils/file_io.py
@@ -1,29 +1,120 @@
-import csv
-import json
-from functools import cache
-import numpy as np
+import sqlite3
 
 
-@cache
-def get_timber_price_table(file_path: str) -> np.ndarray:
-    """Converts the string representation of a timber price table csv to a numpy.ndarray."""
-    table = np.genfromtxt(file_path, delimiter=';', skip_header=1)
-    return table
-
-
-@cache
-def get_renewal_costs_as_dict(file_path: str) -> dict[str, float]:
-    """Returns the csv at :file_path: as a dictionary, where key is the operation name and value ie the cost."""
-    costs = {}
-    with open(file_path, "r", encoding="utf-8") as f:
-        reader = csv.reader(f, delimiter=';')
-        _ = next(reader)  # skip header
-        for row in reader:
-            costs[row[0]] = float(row[1])  # operation: cost
-    return costs
-
-
-@cache
-def get_land_values_as_dict(file_path: str) -> dict:
-    with open(file_path, "r", encoding="utf-8") as f:
-        return json.load(f)
+def create_database_tables(db: sqlite3.Connection):
+    cur = db.cursor()
+    cur.execute(
+        """
+        CREATE TABLE nodes(
+            identifier TEXT,
+            stand TEXT,
+            done_treatment TEXT,
+            treatment_params TEXT,
+            PRIMARY KEY(identifier, stand))
+        """
+    )
+    cur.execute(
+        """
+        CREATE TABLE stands(
+            node TEXT,
+            identifier TEXT,
+            year INTEGER,
+            management_unit_id INTEGER,
+            stand_id INTEGER,
+            area REAL,
+            area_weight REAL,
+            geo_location TEXT,
+            degree_days REAL,
+            owner_category INTEGER,
+            land_use_category INTEGER,
+            soil_peatland_category INTEGER,
+            site_type_category INTEGER,
+            tax_class_reduction INTEGER,
+            tax_class INTEGER,
+            drainage_category INTEGER,
+            drainage_feasibility INTEGER(1),
+            drainage_year INTEGER,
+            fertilization_year INTEGER,
+            soil_surface_preparation_year INTEGER,
+            natural_regeneration_feasibility INTEGER(1),
+            regeneration_area_cleaning_year INTEGER,
+            development_class INTEGER,
+            artificial_regeneration_year INTEGER,
+            young_stand_tending_year INTEGER,
+            pruning_year INTEGER,
+            cutting_year INTEGER,
+            forestry_centre_id INTEGER,
+            forest_management_category REAL,
+            method_of_last_cutting INTEGER,
+            municipality_id INTEGER,
+            dominant_storey_age REAL,
+            area_weight_factors TEXT,
+            fra_category TEXT,
+            land_use_category_detail TEXT,
+            auxiliary_stand INTEGER(1),
+            sea_effect REAL,
+            lake_effect REAL,
+            basal_area REAL,
+            PRIMARY KEY(node, identifier),
+            FOREIGN KEY(node, identifier) REFERENCES nodes(identifier, stand))
+        """
+    )
+    cur.execute(
+        """
+        CREATE TABLE trees(
+            node TEXT,
+            stand TEXT,
+            identifier TEXT,
+            tree_number INTEGER,
+            species INTEGER,
+            breast_height_diameter REAL,
+            height REAL,
+            measured_height REAL,
+            breast_height_age REAL,
+            biological_age REAL,
+            stems_per_ha REAL,
+            origin INTEGER,
+            management_category INTEGER,
+            saw_log_volume_reduction_factor REAL,
+            pruning_year INTEGER,
+            age_when_10cm_diameter_at_breast_height INTEGER,
+            stand_origin_relative_position TEXT,
+            lowest_living_branch_height REAL,
+            tree_category TEXT,
+            storey INTEGER,
+            sapling INTEGER(1),
+            tree_type TEXT,
+            tuhon_ilmiasu TEXT,
+            PRIMARY KEY (node, identifier),
+            FOREIGN KEY (node, stand) REFERENCES nodes(identifier, stand))
+        """
+    )
+    cur.execute(
+        """
+        CREATE TABLE strata(
+            node TEXT,
+            stand TEXT,
+            identifier TEXT,
+            species INTEGER,
+            mean_diameter REAL,
+            mean_height REAL,
+            breast_height_age REAL,
+            biological_age REAL,
+            stems_per_ha REAL,
+            basal_area REAL,
+            origin INTEGER,
+            management_category INTEGER,
+            saw_log_volume_reduction_factor REAL,
+            cutting_year INTEGER,
+            age_when_10cm_diameter_at_breast_height INTEGER,
+            tree_number INTEGER,
+            stand_origin_relative_position TEXT,
+            lowest_living_branch_height REAL,
+            storey INTEGER,
+            sapling_stems_per_ha REAL,
+            sapling_stratum INTEGER(1),
+            number_of_generated_trees INTEGER,
+            PRIMARY KEY (node, identifier),
+            FOREIGN KEY (node, stand) REFERENCES nodes(identifier, stand))
+        """
+    )