+++ CHANGELOG.md
@@ -4,11 +4,29 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## [0.0.7] - 2025-10-22
+## [0.1.3] - 2025-10-31
 
 ### Added
 
 - Added regeneration treatment and planting event
+
+## [0.1.2] - 2025-10-31
+
+### Changed
+
+- Output database columns are now properly typed
+
+## [0.1.1] - 2025-10-27
+
+### Added
+
+- Added operation Soil preparation #MELA2-110
+
+## [0.1.0] - 2025-10-17
+
+### Changed
+
+- Simulation results are now output into an SQLite database
 
 ## [0.0.6] - 2025-10-17